Renamed navigation and scss files
