+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { Link } from "gatsby"
-import Navigation from "../components/navigation"
+import Navigation from "../components/nav"
 import Layout from "../components/layout"
 import SEO from "../components/seo"
