+++ src/sass/nav.scss
@@ -1,4 +1,4 @@
-@import "variables.scss";
+@import "vars.scss";
 
 nav {
   top: 0;
